# public/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Todo CRUD API",
    "description": "API for managing todos.",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/todos": {
      "post": {
        "operationId": "createTodo",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateTodoRequestBody" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Response for status 201",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateTodo201ResponseBody"
                }
              }
            }
          },
          "401": {
            "description": "Response for status 401",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateTodo401ResponseBody"
                }
              }
            }
          },
          "500": {
            "description": "An unknown error occurred, trying again might help.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UnexpectedError" }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "getTodos",
        "responses": {
          "200": {
            "description": "Response for status 200",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetTodos200ResponseBody"
                }
              }
            }
          },
          "500": {
            "description": "An unknown error occurred, trying again might help.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UnexpectedError" }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "updateTodo",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateTodoRequestBody" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Response for status 200",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateTodo200ResponseBody"
                }
              }
            }
          },
          "401": {
            "description": "Response for status 401",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateTodo401ResponseBody"
                }
              }
            }
          },
          "404": {
            "description": "Response for status 404",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateTodo404ResponseBody"
                }
              }
            }
          },
          "500": {
            "description": "An unknown error occurred, trying again might help.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UnexpectedError" }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteTodo",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DeleteTodoRequestBody" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Response for status 200",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteTodo200ResponseBody"
                }
              }
            }
          },
          "401": {
            "description": "Response for status 401",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteTodo401ResponseBody"
                }
              }
            }
          },
          "404": {
            "description": "Response for status 404",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteTodo404ResponseBody"
                }
              }
            }
          },
          "500": {
            "description": "An unknown error occurred, trying again might help.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UnexpectedError" }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CreateTodo201ResponseBody": { "type": "string" },
      "CreateTodo401ResponseBody": { "type": "string" },
      "CreateTodoRequestBody": {
        "type": "object",
        "properties": { "name": { "type": "string" } },
        "required": ["name"],
        "additionalProperties": false
      },
      "DeleteTodo200ResponseBody": { "type": "string" },
      "DeleteTodo401ResponseBody": { "type": "string" },
      "DeleteTodo404ResponseBody": { "type": "string" },
      "DeleteTodoRequestBody": {
        "type": "object",
        "properties": { "id": { "type": "number" } },
        "required": ["id"],
        "additionalProperties": false
      },
      "GetTodos200ResponseBody": {
        "type": "array",
        "items": {
          "type": "object",
          "properties": {
            "id": { "type": "number" },
            "name": { "type": "string" },
            "completed": { "type": "boolean" }
          },
          "required": ["id", "name", "completed"],
          "additionalProperties": false
        }
      },
      "UnexpectedError": {
        "type": "object",
        "properties": { "message": { "type": "string" } },
        "additionalProperties": false
      },
      "UpdateTodo200ResponseBody": { "type": "string" },
      "UpdateTodo401ResponseBody": { "type": "string" },
      "UpdateTodo404ResponseBody": { "type": "string" },
      "UpdateTodoRequestBody": {
        "type": "object",
        "properties": {
          "id": { "type": "number" },
          "completed": { "type": "boolean" }
        },
        "required": ["id", "completed"],
        "additionalProperties": false
      }
    }
  }
}